PI-494 added blueprint model permissions view
